frontend: Project Catalog - Rework of layout
