+++ frontend/packages/core/src/index.tsx
@@ -61,6 +61,7 @@ export { Typography } from "./typography";
 export { default as ClutchApp } from "./AppProvider";
 export { useTheme } from "./AppProvider/themes";
 export { ThemeProvider } from "./Theme";
+export { getDisplayName } from "./utils";
 
 export { css as EMOTION_CSS, keyframes as EMOTION_KEYFRAMES } from "@emotion/react";
