+++ frontend/workflows/projectCatalog/src/details/context.tsx
@@ -3,6 +3,7 @@ import type { clutch as IClutch } from "@clutch-sh/api";
 
 interface ContextProps {
   projectInfo: IClutch.core.project.v1.IProject | null;
+  projectId: string | null;
 }
 
 const ProjectDetailsContext = React.createContext<ContextProps | undefined>(undefined);
