+++ frontend/workflows/projectCatalog/src/details/index.tsx
@@ -1,56 +1,29 @@
 import React from "react";
-import { useParams } from "react-router-dom";
-import type { clutch as IClutch } from "@clutch-sh/api";
-import {
-  checkFeatureEnabled,
-  Grid,
-  IconButton,
-  Link,
-  Popper,
-  PopperItem,
-  QuickLinkGroup,
-  QuickLinksCard,
-  styled,
-  Tooltip,
-  Typography,
-} from "@clutch-sh/core";
+import { Grid, Tooltip, useTheme } from "@clutch-sh/core";
 import CodeOffIcon from "@mui/icons-material/CodeOff";
 import GroupIcon from "@mui/icons-material/Group";
-import SettingsIcon from "@mui/icons-material/Settings";
-import { capitalize, isEmpty } from "lodash";
+import { capitalize } from "lodash";
 
-import type { CatalogDetailsChild, ProjectConfigLink, ProjectDetailsWorkflowProps } from "..";
-
-import { CardType, DynamicCard, MetaCard } from "./card";
-import { ProjectDetailsContext } from "./context";
-import ProjectHeader from "./header";
-import { fetchProjectInfo } from "./helpers";
+import { CardType, DetailCard, DynamicCard, MetaCard } from "./components/card";
 import ProjectInfoCard from "./info";
+import { useMediaQuery } from "@mui/material";
+import { type WorkflowProps, type ProjectCatalogProps, useProjectDetailsContext } from "..";
+import type { ProjectInfoChip } from "./info/chipsRow";
+import CatalogLayout from "./components/layout";
 
-interface QuickLinksAndSettingsProps {
-  linkGroups: QuickLinkGroup[];
-  configLinks?: ProjectConfigLink[];
+export type CatalogDetailsChild = React.ReactElement<DetailCard>;
+
+export interface ProjectConfigLink {
+  title: string;
+  path: string;
+  icon?: React.ReactElement;
 }
 
-const StyledContainer = styled(Grid)({
-  padding: "16px",
-});
-
-const StyledHeadingContainer = styled(Grid)({
-  marginBottom: "24px",
-});
-
-const StyledPopperItem = styled(PopperItem)({
-  "&&&": {
-    height: "auto",
-  },
-  "& span.MuiTypography-root": {
-    padding: "0",
-  },
-  "& a.MuiTypography-root": {
-    padding: "4px 16px",
-  },
-});
+export interface ProjectDetailsWorkflowProps extends WorkflowProps, ProjectCatalogProps {
+  children?: CatalogDetailsChild | CatalogDetailsChild[];
+  chips?: ProjectInfoChip[];
+  configLinks?: ProjectConfigLink[];
+}
 
 const DisabledItem = ({ name }: { name: string }) => (
   <Grid item>
@@ -60,87 +33,13 @@ const DisabledItem = ({ name }: { name: string }) => (
   </Grid>
 );
 
-const QuickLinksAndSettingsBtn = ({ linkGroups, configLinks = [] }: QuickLinksAndSettingsProps) => {
-  const { projectId } = useParams();
-  const anchorRef = React.useRef(null);
-  const [open, setOpen] = React.useState(false);
-  const [links, setLinks] = React.useState<ProjectConfigLink[]>(configLinks);
-
-  React.useEffect(() => {
-    const projectConfigFlag = checkFeatureEnabled({ feature: "projectCatalogSettings" });
-    if (projectConfigFlag) {
-      setLinks([
-        {
-          title: "Project Configuration",
-          path: `/catalog/${projectId}/config`,
-          icon: <SettingsIcon fontSize="small" />,
-        },
-        ...links,
-      ]);
-    }
-  }, []);
-
-  return (
-    <Grid
-      container
-      direction="row"
-      alignItems="center"
-      justifyContent="flex-end"
-      spacing={1}
-      style={{
-        padding: "8px 0px 0px 0px",
-      }}
-    >
-      {!isEmpty(linkGroups) && (
-        <Grid item>
-          <QuickLinksCard linkGroups={linkGroups} />
-        </Grid>
-      )}
-      {links && links.length > 0 && (
-        <Grid item>
-          <IconButton ref={anchorRef} onClick={() => setOpen(o => !o)} size="medium">
-            <SettingsIcon />
-          </IconButton>
-          <Popper
-            open={open}
-            anchorRef={anchorRef}
-            onClickAway={() => setOpen(false)}
-            placement="bottom-end"
-          >
-            {links.map(link => (
-              <StyledPopperItem key={link.title}>
-                <Link href={link.path} target="_self">
-                  <Grid container gap={0.5}>
-                    {link.icon && <Grid item>{link.icon}</Grid>}
-                    <Grid item>
-                      <Typography variant="body2" color="inherit">
-                        {link.title}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-                </Link>
-              </StyledPopperItem>
-            ))}
-          </Popper>
-        </Grid>
-      )}
-    </Grid>
-  );
-};
-
-const Details: React.FC<ProjectDetailsWorkflowProps> = ({
-  children,
-  chips,
-  allowDisabled,
-  configLinks = [],
-}) => {
-  const { projectId } = useParams();
-  const [projectInfo, setProjectInfo] = React.useState<IClutch.core.project.v1.IProject | null>(
-    null
-  );
+const Details = ({ children, chips }: ProjectDetailsWorkflowProps) => {
+  const { projectId, projectInfo } = useProjectDetailsContext() || {};
   const [metaCards, setMetaCards] = React.useState<CatalogDetailsChild[]>([]);
   const [dynamicCards, setDynamicCards] = React.useState<CatalogDetailsChild[]>([]);
-  const projInfo = React.useMemo(() => ({ projectInfo }), [projectInfo]);
+
+  const theme = useTheme();
+  const shrink = useMediaQuery(theme.breakpoints.down("lg"));
 
   React.useEffect(() => {
     if (children) {
@@ -153,16 +52,16 @@ const Details: React.FC<ProjectDetailsWorkflowProps> = ({
 
           switch (type) {
             case CardType.METADATA:
-              tempMetaCards.push(child);
+              tempMetaCards.push(child as CatalogDetailsChild);
               break;
             case CardType.DYNAMIC:
-              tempDynamicCards.push(child);
+              tempDynamicCards.push(child as CatalogDetailsChild);
               break;
             default: {
               if (child.type === DynamicCard) {
-                tempDynamicCards.push(child);
+                tempDynamicCards.push(child as CatalogDetailsChild);
               } else if (child.type === MetaCard) {
-                tempMetaCards.push(child);
+                tempMetaCards.push(child as CatalogDetailsChild);
               }
               // Do nothing, invalid card
             }
@@ -193,70 +92,62 @@ const Details: React.FC<ProjectDetailsWorkflowProps> = ({
   };
 
   return (
-    <ProjectDetailsContext.Provider value={projInfo}>
-      <StyledContainer container direction="row" wrap="nowrap">
-        {/* Column for project details and header */}
-        <Grid container item direction="column" xs={12} sm={12} md={12} lg={12} xl={12}>
-          <Grid container item>
-            <StyledHeadingContainer item xs={6} sm={6} md={7} lg={8} xl={9}>
-              {/* Static Header */}
-              <ProjectHeader
-                title={projectId}
-                routes={[{ title: projectId }]}
-                description={projectInfo?.data?.description as string}
-              />
-            </StyledHeadingContainer>
-            {projectInfo && (
-              <Grid container item xs={12} sm={12} md={5} lg={4} xl={3} spacing={2}>
-                <QuickLinksAndSettingsBtn
-                  linkGroups={(projectInfo.linkGroups as QuickLinkGroup[]) || []}
-                  configLinks={configLinks ?? []}
-                />
-              </Grid>
-            )}
-          </Grid>
-          <Grid container item direction="row" spacing={2}>
-            <Grid container item xs={12} sm={12} md={5} lg={4} xl={3} spacing={2}>
-              <Grid item xs={12}>
-                {/* Static Info Card */}
-                <MetaCard
-                  title={getOwner(projectInfo?.owners ?? []) || projectId}
-                  titleIcon={<GroupIcon />}
-                  fetchDataFn={() => fetchProjectInfo(projectId, allowDisabled)}
-                  onSuccess={(data: unknown) =>
-                    setProjectInfo(data as IClutch.core.project.v1.IProject)
-                  }
-                  autoReload
-                  loadingIndicator={false}
-                  endAdornment={
-                    projectInfo?.data?.disabled ? <DisabledItem name={projectId} /> : null
-                  }
-                >
-                  {projectInfo && <ProjectInfoCard projectData={projectInfo} addtlChips={chips} />}
-                </MetaCard>
-              </Grid>
-              {/* Custom Meta Cards */}
-              {metaCards.length > 0 &&
-                metaCards.map(card => (
-                  <Grid item xs={12}>
-                    {card}
-                  </Grid>
-                ))}
-            </Grid>
-            <Grid container item xs={12} sm={12} md={7} lg={8} xl={9} spacing={1}>
-              {/* Custom Dynamic Cards */}
-              {dynamicCards.length > 0 &&
-                dynamicCards.map(card => (
-                  <Grid item xs={12} key={`dynamic-${card.key}`}>
-                    {card}
-                  </Grid>
-                ))}
-            </Grid>
-          </Grid>
+    <>
+      <Grid
+        container
+        item
+        direction={shrink ? "row" : "column"}
+        xs={12}
+        sm={12}
+        md={5}
+        lg={4}
+        xl={3}
+        spacing={2}
+        flexWrap="nowrap"
+      >
+        <Grid item>
+          {projectInfo && (
+          <MetaCard
+            title={getOwner(projectInfo?.owners ?? []) || projectInfo?.name}
+            titleIcon={<GroupIcon />}
+            loadingIndicator={false}
+            endAdornment={projectInfo?.data?.disabled ? <DisabledItem name={projectInfo?.name ?? projectId ?? ""} /> : null}
+          >
+            {projectInfo && <ProjectInfoCard projectData={projectInfo} addtlChips={chips} />}
+          </MetaCard>
+
+          )}
         </Grid>
-      </StyledContainer>
-    </ProjectDetailsContext.Provider>
+        {metaCards.length > 0 && metaCards.map(card => <Grid item>{card}</Grid>)}
+      </Grid>
+      <Grid
+        container
+        item
+        direction="column"
+        xs={12}
+        sm={12}
+        md={7}
+        lg={8}
+        xl={9}
+        spacing={2}
+        flexWrap="nowrap"
+      >
+        {dynamicCards.length > 0 && dynamicCards.map(card => <Grid item>{card}</Grid>)}
+      </Grid>
+    </>
   );
 };
 
-export default Details;
+const CatalogDetailsPage = ({
+  allowDisabled,
+  configLinks,
+  ...props
+}: ProjectDetailsWorkflowProps) => {
+  return (
+    <CatalogLayout allowDisabled={allowDisabled} configLinks={configLinks ?? []}>
+      <Details allowDisabled={allowDisabled} {...props} />
+    </CatalogLayout>
+  );
+};
+
+export default CatalogDetailsPage;
