+++ frontend/workflows/projectCatalog/src/helpers/index.ts
@@ -1,0 +1,3 @@
+export { default as LanguageIcon } from "./language-icon";
+export { default as LastEvent } from "./last-event";
+export { default as LinkText } from "./link-text";
