+++ frontend/workflows/projectCatalog/src/helpers/link-text.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { Link, Typography } from "@clutch-sh/core";
+import * as _ from "lodash";
+
+const LinkText = ({ text, link }: { text: string; link?: string }) => {
+  const returnText = <Typography variant="body2">{text}</Typography>;
+
+  if (link && text) {
+    return (
+      <Link href={link} whiteSpace="nowrap">
+        {returnText}
+      </Link>
+    );
+  }
+
+  return returnText;
+};
+
+export default LinkText;
